+++ frontend/src/base/home/forms/AlertPlugin.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+    import { onMount, onDestroy } from 'svelte';
+    
     import { CloseCircleSolid, FloppyDiskSolid } from 'flowbite-svelte-icons';
     import {
         Heading, Button, Input, Label, Spinner, Tooltip, Helper,
@@ -19,6 +21,7 @@
         classModalInput, classModalHelp, classSpinnerDiv, classModalBody, classModalDialog,
     } from '../../Style.js';
 
+    let componentRoot;
     let {
         open = $bindable(false),
         successMsg = $bindable(''),
@@ -38,6 +41,8 @@
     let url = $derived(actionNew ? urlBase : urlExisting);
     let title = $derived(actionNew ? t('alerts.plugin.new') : t('alerts.plugin.edit'));
     let formWarningMsgs: string[] = $state([]);
+    let pressedKeyAlt = $state(false);
+    let pressedKeyS = $state(false);
 
     let form = $state({
         name: {value: '', color: inputBaseColor, required: true, regex: /^.{1,100}/},
@@ -98,8 +103,64 @@
         }
         fetchBuild();
     })
+
+    // ALT+S quick-save
+    function handleKeyUp(e: KeyboardEvent) {
+        switch (e.key) {
+        case 'Alt':
+            pressedKeyAlt = false;
+            break;
+        case 's':
+        case 'S':
+            pressedKeyS = false;
+            break;
+        default:
+            return;
+        }
+    }
+    
+    function handleKeyDown(e: KeyboardEvent) {
+        switch (e.key) {
+        case 'Alt':
+            pressedKeyAlt = true;
+            break;
+        case 's':
+        case 'S':
+            pressedKeyS = true;
+            break;
+        default:
+            return;
+        }
+    }
+
+    $effect(() => {
+        if (open && componentRoot) {
+            componentRoot.focus();
+        }
+    });
+
+    $effect(() => {
+        if (pressedKeyAlt && pressedKeyS) {
+            submitForm();
+        }
+    });
+
+    onMount(() => {
+        if (componentRoot) {
+            componentRoot.addEventListener('keyup', handleKeyUp);
+            componentRoot.addEventListener('keydown', handleKeyDown);
+        }
+    });
+
+    onDestroy(()=>{
+        if (componentRoot) {
+            componentRoot.removeEventListener('keyup', handleKeyUp);
+            componentRoot.removeEventListener('keydown', handleKeyDown);
+        }
+    });
 </script>
 
+<div bind:this={componentRoot} tabindex="-1" class="inline-block">
 <Modal bind:open={open} size="lg" autoclose={false} placement="top-center"
     backdropClass={classModalBackdrop} bodyClass={classModalBody} dialogClass={classModalDialog}>
     <Heading tag="h2">{title}</Heading>
@@ -130,3 +191,4 @@
         </div>
     {/if}
 </Modal>
+</div>