+++ frontend/src/base/home/forms/Credentials.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+    import { onMount, onDestroy } from 'svelte';
+    
     import { CloseCircleSolid, FloppyDiskSolid } from 'flowbite-svelte-icons';
     import {
         Heading, Button, Input, Label, Spinner, Tooltip, Helper,
@@ -24,6 +26,7 @@
         classModalBody, classModalDialog,
     } from '../../Style.js';
 
+    let componentRoot;
     let {
         open = $bindable(false),
         successMsg = $bindable(''),
@@ -44,6 +47,9 @@
     let url = $derived(actionNew ? `credentials/${kind}` : urlExisting);
     let title = $derived(actionNew ? t('creds.new') : t('creds.edit'));
     let formWarningMsgs: string[] = $state([]);
+    let pressedKeyAlt = $state(false);
+    let pressedKeyS = $state(false);
+
     let existing: credentialsUserType|credentialsSharedType = $state({
         id: 0,
         name: '',
@@ -159,8 +165,64 @@
         }
         setTimeout(fetchBuild, 500);  // wait to fetch version
     })
+
+    // ALT+S quick-save
+    function handleKeyUp(e: KeyboardEvent) {
+        switch (e.key) {
+        case 'Alt':
+            pressedKeyAlt = false;
+            break;
+        case 's':
+        case 'S':
+            pressedKeyS = false;
+            break;
+        default:
+            return;
+        }
+    }
+    
+    function handleKeyDown(e: KeyboardEvent) {
+        switch (e.key) {
+        case 'Alt':
+            pressedKeyAlt = true;
+            break;
+        case 's':
+        case 'S':
+            pressedKeyS = true;
+            break;
+        default:
+            return;
+        }
+    }
+
+    $effect(() => {
+        if (open && componentRoot) {
+            componentRoot.focus();
+        }
+    });
+
+    $effect(() => {
+        if (pressedKeyAlt && pressedKeyS) {
+            submitForm();
+        }
+    });
+
+    onMount(() => {
+        if (componentRoot) {
+            componentRoot.addEventListener('keyup', handleKeyUp);
+            componentRoot.addEventListener('keydown', handleKeyDown);
+        }
+    });
+
+    onDestroy(()=>{
+        if (componentRoot) {
+            componentRoot.removeEventListener('keyup', handleKeyUp);
+            componentRoot.removeEventListener('keydown', handleKeyDown);
+        }
+    });
 </script>
 
+<div bind:this={componentRoot} tabindex="-1" class="inline-block">
 <Modal bind:open={open} size="lg" autoclose={false} placement="top-center"
     backdropClass={classModalBackdrop} bodyClass={classModalBody} dialogClass={classModalDialog}>
     <Heading tag="h2">{title}</Heading>
@@ -253,3 +315,4 @@
         </div>
     {/if}
 </Modal>
+</div>