+++ frontend/src/base/home/forms/Job.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { onMount } from 'svelte';
+    import { onMount, onDestroy } from 'svelte';
 
     import {
         FolderSolid, FileSolid, CloseCircleSolid, TrashBinSolid, FloppyDiskSolid, CirclePlusSolid,
@@ -29,6 +29,7 @@
         classModalDialog, classModalBody, classSpoilerPad,
     } from '../../Style.js';
 
+    let componentRoot;
     let {
         open = $bindable(false),
         successMsg = $bindable(''),
@@ -48,6 +49,8 @@
     let url = $derived(actionNew ? 'job' : urlExisting);
     let title = $derived(actionNew ? t('jobs.new') : t('jobs.edit'));
     let formWarningMsgs: string[] = $state([]);
+    let pressedKeyAlt = $state(false);
+    let pressedKeyS = $state(false);
 
     let form = $state({
         name: {value: '', color: inputBaseColor, required: true},
@@ -355,17 +358,70 @@
         }
     }
 
+    // ALT+S quick-save
+    function handleKeyUp(e: KeyboardEvent) {
+        switch (e.key) {
+        case 'Alt':
+            pressedKeyAlt = false;
+            break;
+        case 's':
+        case 'S':
+            pressedKeyS = false;
+            break;
+        default:
+            return;
+        }
+    }
+    
+    function handleKeyDown(e: KeyboardEvent) {
+        switch (e.key) {
+        case 'Alt':
+            pressedKeyAlt = true;
+            break;
+        case 's':
+        case 'S':
+            pressedKeyS = true;
+            break;
+        default:
+            return;
+        }
+    }
+
+    $effect(() => {
+        if (open && componentRoot) {
+            componentRoot.focus();
+        }
+    });
+
+    $effect(() => {
+        if (pressedKeyAlt && pressedKeyS) {
+            submitForm();
+        }
+    });
+
     onMount(() => {
+        if (componentRoot) {
+            componentRoot.addEventListener('keyup', handleKeyUp);
+            componentRoot.addEventListener('keydown', handleKeyDown);
+        }
         updateCreateGitRepo();
-    })
+    });
+
+    onDestroy(()=>{
+        if (componentRoot) {
+            componentRoot.removeEventListener('keyup', handleKeyUp);
+            componentRoot.removeEventListener('keydown', handleKeyDown);
+        }
+    });
 
     $effect(() => {
         if (isSet(form.repository.value)) {
             updateCreateGitRepo();
         }
-    })
+    });
 </script>
 
+<div bind:this={componentRoot} tabindex="-1" class="inline-block">
 <Modal bind:open={open} size="lg" autoclose={false} placement="top-center"
     backdropClass={classModalBackdrop} dialogClass={classModalDialog} bodyClass={classModalBody}>
     <Heading tag="h2">{title}</Heading>
@@ -691,3 +747,4 @@
         </div>
     {/if}
 </Modal>
+</div>